[SPMD] Split `test_ops.py` into multiple files (#312)
