+++ test/spmd/tensor/test_matrix_ops.py
@@ -2,10 +2,10 @@
 import torch
 from torch.testing._internal.common_utils import run_tests
 from ..test_utils import DistTensorTestBase, with_comms
-from spmd import distribute_tensor, DeviceMesh, Tensor, Shard, Replicate
+from spmd import distribute_tensor, DeviceMesh, Shard, Replicate
 
 
-class DistTensorOpsTest(DistTensorTestBase):
+class DistMatrixOpsTest(DistTensorTestBase):
     @with_comms
     def test_addmm(self):
         device_mesh = DeviceMesh(self.device_type, list(range(self.world_size)))
@@ -54,18 +54,6 @@ class DistTensorOpsTest(DistTensorTestBase):
         # dist_res.sum().backward()
 
     @with_comms
-    def test_sum(self):
-        device_mesh = DeviceMesh(self.device_type, list(range(self.world_size)))
-        shard_spec = [Shard(0)]
-        replica_spec = [Replicate()]
-
-        tensor_to_sum = torch.randn(12, 8)
-        sumed_tensor = tensor_to_sum.sum()
-        mat1 = distribute_tensor(tensor_to_sum, device_mesh, shard_spec)
-        dt_sum = mat1.sum()
-        self.assertEqual(dt_sum.local_tensor(), sumed_tensor)
-
-    @with_comms
     def test_t(self):
         device_mesh = DeviceMesh(self.device_type, list(range(self.world_size)))
         shard_spec = [Shard(0)]
@@ -79,28 +67,6 @@ class DistTensorOpsTest(DistTensorTestBase):
         tranposed_mat2 = tranposed_mat.t()
         self.assertEqual(tranposed_mat2.size(), torch.Size([12, 8]))
         self.assertEqual(tranposed_mat2.placements, shard_spec)
-
-    @with_comms
-    def test_detach(self):
-        device_mesh = DeviceMesh(self.device_type, list(range(self.world_size)))
-        shard_spec = [Shard(0)]
-
-        tensor_to_detach = torch.randn(12, 8, requires_grad=True)
-        mat = distribute_tensor(tensor_to_detach, device_mesh, shard_spec)
-        detached_mat = mat.detach()
-        self.assertFalse(detached_mat is mat)
-
-    @with_comms
-    def test_ones_like(self):
-        device_mesh = DeviceMesh(self.device_type, list(range(self.world_size)))
-        shard_spec = [Shard(0)]
-        replica_spec = [Replicate()]
-
-        input_tensor = torch.randn(4, 8, requires_grad=True)
-        dist_tensor = Tensor.from_local(input_tensor, device_mesh, shard_spec)
-        ones_like_dt = torch.ones_like(dist_tensor)
-        ones_expected = torch.ones(4, 8)
-        self.assertEqual(ones_expected, ones_like_dt.local_tensor())
 
 
 if __name__ == "__main__":
